combination of workspace and stageplanner complete
